Option to disable auto association of revision model with user model
Currently revision model tries to establish association with user
model during model creation. However, this does not work if the
user model is not defined at that moment. This would be the case if
we want to enable paper trail for the user model itself. In such
cases it is better to disable auto associating with user model,
and later associate explicityly using revision.associate function call.